Update: Register and admin auth

+++ app/Http/Controllers/PageController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\DB;
+use App\User;
+use Auth;
 
 class PageController extends Controller
 {
@@ -17,20 +19,68 @@ class PageController extends Controller
         } else {
             return redirect()->route('home');
         }
+        return view('cv');
     }
 
     public function admin()
     {
-        if (Session::get('isLoggedIn') == true) {
-            $karyawan = DB:: table('datakaryawan')->get();
-            
+        if(Auth::user()->role != 'admin')
+        {
+            return redirect()->route('home');
+        }
+        else
+        {
+            if (Session::get('isLoggedIn') == true) {
+                $karyawan = DB:: table('datakaryawan')->get();
+                
 
-            return view('admin',['karyawan' => $karyawan]);
+                return view('admin',['karyawan' => $karyawan]);
+                
+            } else {
+                return redirect()->route('home');
+            }
+        }
+    }
+
+    //view untuk register
+    public function register(Request $request)
+    {
+        if(Auth::user()->role != 'admin')
+        {
+            return redirect()->route('home');
+        }
+
+        if (Session::get('isLoggedIn') == true) {
+            //  $users = DB:: table('users')->get();
+            //  DB::table('users')->insert([
+            
+            //     'email' => $request->email,
+            //     'password' => $request->password,
+            //     'nama' => $request->nama,
+            //     'role'=> 'member' 
+            //]);
+            return view('/register');
             
         } else {
             return redirect()->route('home');
         }
     }
+
+    
+    // method untuk insert data ke table users
+    public function userstore(Request $request)
+        {
+            // insert data ke table user
+            DB::table('users')->insert([
+            
+                'email' => $request->email,
+                'password' => bcrypt($request->password),
+                'nama' => $request->nama,
+                'role'=> 'member' 
+            ]);
+            // alihkan halaman ke halaman pegawai
+            return redirect('/admin');    
+        }
 
         // method untuk insert data ke table datakaryawan
      public function store(Request $request)
@@ -46,9 +96,14 @@ class PageController extends Controller
             return redirect('/admin');
     
         }
-
-     public function edit($id)
+    
+        //untuk view edit
+        public function edit($id)
         {
+            if(Auth::user()->role != 'admin')
+            {
+                return redirect()->route('home');
+            }
             // mengambil data pegawai berdasarkan id yang dipilih
             $karyawan = DB::table('datakaryawan')->get()->where('no',$id);
             // passing data pegawai yang didapat ke view edit.blade.php
@@ -83,6 +138,10 @@ class PageController extends Controller
         // method untuk hapus data pegawai
         public function hapus($id)
         {
+            if(Auth::user()->role != 'admin')
+            {
+                return redirect()->route('home');
+            }
             // menghapus data pegawai berdasarkan id yang dipilih
             DB::table('datakaryawan')->where('no',$id)->delete();
